Revert "Add react-router & more cleanup"
This reverts commit 5218940769a14048e715b5769049e11cc8f1991e.

+++ server/src/graphql/typeDefs.ts
@@ -10,7 +10,7 @@ export const typeDefs = gql`
     numOfGuests: Int!
     numOfBeds: Int!
     numOfBaths: Int!
-    rating: Float!
+    rating: Int!
   }
 
   type Query {
